refactoring
This is working towards
https://github.com/ossf/fuzz-introspector/issues/835 by making the code
more modular.
