+++ src/fuzz_introspector/html_helpers.py
@@ -15,6 +15,7 @@
 
 from typing import (
     Any,
+    Dict,
     List,
     Optional,
     Tuple,
@@ -22,10 +23,13 @@ from typing import (
 
 import os
 import bs4
+import logging
 from enum import Enum
 
-from fuzz_introspector import utils
+from fuzz_introspector import utils, constants
 from fuzz_introspector.datatypes import fuzzer_profile, project_profile
+
+logger = logging.getLogger(name=__name__)
 
 
 class HTML_HEADING(Enum):
@@ -385,6 +389,117 @@ def create_conclusions_box(conclusions: List[HTMLConclusion]) -> str:
     # it shuold be there. Verify.
     html_string += "</div>"
     return html_string
+
+
+def create_calltree_color_distribution_table(color_list: List[str]) -> str:
+    html_string = ""
+    color_dictionary: Dict[str, int] = {}
+    for color in color_list:
+        color_dictionary[color] = color_dictionary.get(color, 0) + 1
+
+    html_string += ("<p>The distribution of callsites in terms of coloring is")
+    html_string += ("<table><tr>"
+                    "<th style=\"text-align: left;\">Color</th>"
+                    "<th style=\"text-align: left;\">Runtime hitcount</th>"
+                    "<th style=\"text-align: left;\">Callsite count</th>"
+                    "<th style=\"text-align: left;\">Percentage</th>"
+                    "</tr>")
+    for _min, _max, color, rgb_code in constants.COLOR_CONSTANTS:
+        html_string += (f"<tr><td style=\"color:{color}; "
+                        f"text-shadow: -1px 0 black, 0 1px black, "
+                        f"1px 0 black, 0 -1px black;\"><b>{color}</b></td>")
+        if _max == 1:
+            interval = "0"
+        elif _max > 1000:
+            interval = f"{_min}+"
+        else:
+            interval = f"[{_min}:{_max-1}]"
+        html_string += f"<td>{interval}</td>"
+        cover_count = color_dictionary.get(color, 0)
+        html_string += f"<td>{cover_count}</td>"
+        if len(color_list) > 0:
+            f1 = float(cover_count)
+            f2 = float(len(color_list))
+            percentage_c = (f1 / f2) * 100.0
+        else:
+            percentage_c = 0.0
+        percentage_s = str(percentage_c)[0:4]
+        html_string += f"<td>{percentage_s}%</td>"
+        html_string += "</tr>"
+
+    # Add a row with total amount of callsites
+    html_string += f"<tr><td>All colors</td><td>{len(color_list)}</td><td>100</td></tr>"
+    html_string += "</table>"
+    html_string += "</p>"
+    return html_string
+
+
+def create_horisontal_calltree_image(image_name: str,
+                                     profile: fuzzer_profile.FuzzerProfile,
+                                     dump_files: bool) -> List[str]:
+    """
+    Creates a horisontal image of the calltree. The height is fixed and
+    each element on the x-axis shows a node in the calltree in the form
+    of a rectangle. The rectangle is red if not visited and green if visited.
+    """
+    try:
+        import matplotlib.pyplot as plt
+        from matplotlib.patches import Rectangle
+    except ModuleNotFoundError:
+        # It's useful to avoid this in CIFuzz because building the fuzzers with
+        # matplotlib costs a lot of time (10 minutes) in the CI, which we prefer
+        # to avoid.
+        logger.info("Could not import matplotlib. No bitmaps are created")
+        return []
+
+    logger.info(f"Creating image {image_name}")
+
+    # Get the callsites of the profile as a list of colors.
+    color_list: List[str] = [cs.cov_color for cs in profile.get_callsites()]
+    logger.info(f"- extracted the callsites ({len(color_list)} nodes)")
+
+    # Show one read rectangle if the list is empty. An alternative is
+    # to not include the image at all.
+    if len(color_list) == 0:
+        color_list = ['red']
+
+    # Create a plot
+    fig, ax = plt.subplots()
+    ax.clear()
+    fig.set_size_inches(15, 2.5)
+    ax.plot()
+
+    # Create our rectangles
+    curr_x = 0.0
+    curr_size = 1.0
+    curr_color = color_list[0]
+    for i in range(1, len(color_list)):
+        if curr_color == color_list[i]:
+            curr_size += 1.0
+        else:
+            ax.add_patch(
+                Rectangle((curr_x, 0.0), curr_size, 1.0, color=curr_color))
+
+            # Start next color area
+            curr_x += curr_size
+            curr_color = color_list[i]
+            curr_size = 1.0
+    # Plot the last case
+    ax.add_patch(Rectangle((curr_x, 0.0), curr_size, 1.0, color=curr_color))
+    logger.info("- iterated over color list")
+
+    # Save the image
+    if dump_files:
+        logger.info("- saving image")
+        ax.set_yticklabels([])
+        ax.set_yticks([])
+        xlabel = ax.set_xlabel("Callsite index")
+
+        plt.title(image_name.replace(".png", "").replace("_colormap", ""))
+        fig.tight_layout()
+        fig.savefig(image_name, bbox_extra_artists=[xlabel])
+        logger.info("- image saved")
+    return color_list
 
 
 def prettify_html(html_doc: str) -> str:
